blog/requirements.txt
asgiref==3.4.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
bs4==0.0.1
cffi==1.14.6
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
distlib==0.3.2
Django==3.2.7
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-mdeditor==0.1.18
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
ipython==7.26.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
jieba==0.42.1
Jinja2==3.0.1
jupyter-client==7.0.1
jupyter-core==4.7.1
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mysql-connector-python==8.0.26
nest-asyncio==1.5.1
numpy==1.21.1
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.2.0
prompt-toolkit==3.0.20
pycparser==2.20
Pygments==2.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pyzmq==22.2.1
simpleui==4.0.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tornado==6.1
traitlets==5.0.5
virtualenv==20.7.0
wcwidth==0.2.5
Werkzeug==2.0.1
xlwt==1.3.0
